Counters controller specs (#3)
* Add specs for counters controller

Add specs for the #show method. Render a counter when it doesn't have
any events, initialized to 0. When a counter has prior events, apply
these events before returning the counter current value.

* Handle reset

Add spec coverage for the #destroy controller method, which reset the
counter and returns its new value (0).

* Handle update controller action

Add better cast to integer in the counter model to ensure we can pass
`nil` as a parameter (instead of the optional value) and still have it
default to `1` in the command.