Add Account model and HTTP endpoint (#6)
* Add account model

* Handle deposit and withdrawal

Convert these two operations into an event sourcing approach. Handle
commands and validate the amount on each operation.

* More specs for Account

Better errors when handling a command. Add specs for withdraw and
deposit operations.

* HTTP Controller for Account

Add JSON:API as a MIME type. Include the fast_jsonapi gem from Netflix
to the bundle. Expose the account as an HTTP GET endpoint.

* Add transactions controller

Allow retrieval of all transactions (= events) associated with an
account through a JSON:API response. Allow creation of new transactions
which execute (= handle) a command on the model.

* Controller specs

Ensure we return with a proper 4xx error when the request is invalid,
and 201 if we create the transaction.